New Design
It has been brought to my attention that the old black and white design
was a bit flat and boring. Plus going over the data for putting a squad
together has lead me to a few design problems. The new design aims to
eleminate those problems and jazz up the design with some color forms.